refactor: rename and generalize dependency-bot precheck script (DBT-1.1)
Renames check-dependabot-triage.ts to check-dependency-bot-triage.ts and
widens PR discovery to query both app/dependabot and app/renovate authors
per repo, merging and deduping results by PR number. Adds bot and
triagedCommitSha fields to the state entry schema so a PR is re-flagged for
triage when its live headRefOid diverges from the last-triaged commit,
mirroring the main review pipeline's staleness check. State path moves to
state/dependency-bot-reviews.json.
